# backend/openapi-schema.yaml
openapi: 3.0.3
info:
  title: Shava API
  version: 1.0.0
  description: 'Public REST API for the Shava project: places, reviews, ratings, users
    (JWT auth), articles, gamification.'
paths:
...
  /api/v1/places/{id}/:
    get:
      operationId: places_retrieve
      description: Author or admin can edit a place.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - places
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceUpdate'
          description: ''
    put:
      operationId: places_update
      description: Author or admin can edit a place.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - places
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PlaceUpdateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PlaceUpdateRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceUpdateRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceUpdate'
          description: ''
    patch:
      operationId: places_partial_update
      description: Author or admin can edit a place.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - places
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedPlaceUpdateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedPlaceUpdateRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedPlaceUpdateRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceUpdate'
          description: ''
...
  /api/v1/places/create-place/:
    post:
      operationId: places_create_place_create
      description: Authenticated users submit a new place; always created on moderation.
      tags:
      - places
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PlaceCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PlaceCreateRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceCreateRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceCreate'
          description: ''
...
  /api/v1/places/place/{id}/:
    get:
      operationId: places_place_retrieve
      description: Public detail view; non-public statuses only visible to author
        or staff.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - places
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceDetail'
          description: ''
  /api/v1/places/ratings/:
    get:
      operationId: places_ratings_list
      description: A user's own ratings (for management UI).
      parameters:
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - places
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPlaceRatingList'
          description: ''
    post:
      operationId: places_ratings_create
      description: A user's own ratings (for management UI).
      tags:
      - places
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceRatingRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PlaceRatingRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PlaceRatingRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceRating'
          description: ''
  /api/v1/places/ratings/{id}/:
    get:
      operationId: places_ratings_retrieve
      description: A user's own ratings (for management UI).
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Place Rating.
        required: true
      tags:
      - places
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceRating'
          description: ''
    put:
      operationId: places_ratings_update
      description: A user's own ratings (for management UI).
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Place Rating.
        required: true
      tags:
      - places
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceRatingRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PlaceRatingRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PlaceRatingRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceRating'
          description: ''
    patch:
      operationId: places_ratings_partial_update
      description: A user's own ratings (for management UI).
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Place Rating.
        required: true
      tags:
      - places
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedPlaceRatingRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedPlaceRatingRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedPlaceRatingRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceRating'
          description: ''
    delete:
      operationId: places_ratings_destroy
      description: A user's own ratings (for management UI).
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Place Rating.
        required: true
      tags:
      - places
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
...
components:
  schemas:
...
    DistrictEnum:
      enum:
      - Unknown
      - Dnipro
      - Desna
      - Solomyanka
      - Shevchenko
      - Pechersk
      - Obolon
      - Podil
      - Svyatoshyn
      - Holosiiv
      - Darnytsia
      type: string
      description: |-
        * `Unknown` - Unknown
        * `Dnipro` - Дніпровський
        * `Desna` - Деснянський
        * `Solomyanka` - Солом'янський
        * `Shevchenko` - Шевченківський
        * `Pechersk` - Печерський
        * `Obolon` - Оболонський
        * `Podil` - Подільський
        * `Svyatoshyn` - Святошинський
        * `Holosiiv` - Голосіївський
        * `Darnytsia` - Дарницький
...
    PaginatedPlaceRatingList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/PlaceRating'
...
    PatchedPlaceRatingRequest:
      type: object
      properties:
        place:
          type: integer
        rating:
          type: string
          format: decimal
          pattern: ^-?\d{0,2}(?:\.\d{0,2})?$
...
    PatchedPlaceUpdateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        city:
          type: string
          minLength: 1
          maxLength: 100
        district:
          $ref: '#/components/schemas/DistrictEnum'
        address:
          type: string
          minLength: 1
        delivery:
          type: boolean
        latitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
          nullable: true
        longitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
          nullable: true
        description:
          type: string
          nullable: true
        main_image:
          type: string
          format: binary
        additional_images:
          type: string
          format: binary
          nullable: true
        website:
          type: string
          format: uri
          nullable: true
          maxLength: 200
        opening_hours:
          type: string
          nullable: true
          maxLength: 100
        status:
          $ref: '#/components/schemas/StatusEnum'
...
    PlaceCreate:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
        city:
          type: string
          maxLength: 100
        address:
          type: string
          maxLength: 300
        delivery:
          type: boolean
        latitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
        longitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
        description:
          type: string
          nullable: true
        main_image:
          type: string
          format: uri
        additional_images:
          type: string
          format: uri
          nullable: true
        website:
          type: string
          format: uri
          nullable: true
          maxLength: 200
        opening_hours:
          type: string
          nullable: true
          maxLength: 100
      required:
      - address
      - city
      - latitude
      - longitude
      - main_image
    PlaceCreateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        city:
          type: string
          minLength: 1
          maxLength: 100
        address:
          type: string
          minLength: 1
          maxLength: 300
        delivery:
          type: boolean
        latitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
        longitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
        description:
          type: string
          nullable: true
        main_image:
          type: string
          format: binary
        additional_images:
          type: string
          format: binary
          nullable: true
        website:
          type: string
          format: uri
          nullable: true
          maxLength: 200
        opening_hours:
          type: string
          nullable: true
          maxLength: 100
      required:
      - address
      - city
      - latitude
      - longitude
      - main_image
    PlaceDetail:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 200
        city:
          type: string
          maxLength: 100
        city_ref:
          type: integer
          nullable: true
          description: Optional reference to the canonical City row. The free-text
            `city` field above is preserved for legacy/back-compat reads.
        district:
          $ref: '#/components/schemas/DistrictEnum'
        address:
          type: string
          maxLength: 300
        delivery:
          type: boolean
        latitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
          nullable: true
        longitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
          nullable: true
        description:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/StatusEnum'
        rating:
          type: string
          format: decimal
          pattern: ^-?\d{0,2}(?:\.\d{0,2})?$
          description: Average rating calculated from reviews
        stars:
          type: string
          readOnly: true
        ratings_count:
          type: string
          readOnly: true
        main_image:
          type: string
          format: uri
        main_image_thumbnails:
          type: string
          readOnly: true
        additional_images:
          type: string
          format: uri
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        website:
          type: string
          format: uri
          nullable: true
          maxLength: 200
        opening_hours:
          type: string
          nullable: true
          maxLength: 100
        is_featured:
          type: boolean
        author:
          type: integer
          nullable: true
        moderated_by:
          type: integer
          nullable: true
        moderation_reason:
          type: string
          nullable: true
        moderated_at:
          type: string
          format: date-time
          nullable: true
        google_maps_url:
          type: string
          readOnly: true
        average_rating:
          type: string
          readOnly: true
        reviews_count:
          type: string
          readOnly: true
        ratings:
          type: array
          items:
            $ref: '#/components/schemas/PlaceRating'
          readOnly: true
        menu:
          type: string
          readOnly: true
      required:
      - address
      - average_rating
      - created_at
      - google_maps_url
      - id
      - main_image
      - main_image_thumbnails
      - menu
      - ratings
      - ratings_count
      - reviews_count
      - stars
      - updated_at
    PlaceRating:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        place:
          type: integer
        rating:
          type: string
          format: decimal
          pattern: ^-?\d{0,2}(?:\.\d{0,2})?$
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - id
      - place
      - rating
      - updated_at
    PlaceRatingRequest:
      type: object
      properties:
        place:
          type: integer
        rating:
          type: string
          format: decimal
          pattern: ^-?\d{0,2}(?:\.\d{0,2})?$
      required:
      - place
      - rating
    PlaceUpdate:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
        city:
          type: string
          maxLength: 100
        district:
          $ref: '#/components/schemas/DistrictEnum'
        address:
          type: string
        delivery:
          type: boolean
        latitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
          nullable: true
        longitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
          nullable: true
        description:
          type: string
          nullable: true
        main_image:
          type: string
          format: uri
        additional_images:
          type: string
          format: uri
          nullable: true
        website:
          type: string
          format: uri
          nullable: true
          maxLength: 200
        opening_hours:
          type: string
          nullable: true
          maxLength: 100
        status:
          $ref: '#/components/schemas/StatusEnum'
    PlaceUpdateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        city:
          type: string
          minLength: 1
          maxLength: 100
        district:
          $ref: '#/components/schemas/DistrictEnum'
        address:
          type: string
          minLength: 1
        delivery:
          type: boolean
        latitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
          nullable: true
        longitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
          nullable: true
        description:
          type: string
          nullable: true
        main_image:
          type: string
          format: binary
        additional_images:
          type: string
          format: binary
          nullable: true
        website:
          type: string
          format: uri
          nullable: true
          maxLength: 200
        opening_hours:
          type: string
          nullable: true
          maxLength: 100
        status:
          $ref: '#/components/schemas/StatusEnum'
...
    StatusEnum:
      enum:
      - Active
      - Inactive
      - On_moderation
      - Approved
      - Rejected
      - Closed
      - Archived
      type: string
      description: |-
        * `Active` - Активний
        * `Inactive` - Неактивний
        * `On_moderation` - На модерації
        * `Approved` - Схвалено
        * `Rejected` - Відхилено
        * `Closed` - Закритий
        * `Archived` - Архівний
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT